requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
azure-identity
azure-keyvault-secrets
openai